Trying out someone elses mnist using Micrograd
As the title says, I'm trying out someone elses mnist to figure out why mine is broken.
I am using a different Loss Calculation which is giving me an abysmal loss.
I will try to figure out why this is happening. Also my imports are a bit broken
because the Engine Folder has the same name as Engine.py I have fixed it
for now but it is still a little bit of a quick patch.